Reorganize code for readability; Separate files for classes.
